[WIP] ABT - Automatic Build Tool
Thanks @Furrydaus and @akhilnarang (resurrection remix) for the idea

-- ATTENTION --
1) Don't rename the program.
2) The sh files have to be in the same folder as the program

TODO: Improve the sentences. I do not speak English very well, sorry. I look forward to better translations - Waiting for some kind soul
TODO: Add support to the translations if possible - Being processed

Tell me if I can improve the tool in some way, thanks.

What does the tool do:
The tool installs the necessary dependencies, and download the source code needed to build the rom.
It allows finally to build the rom, and provides an executable file from bash that guides the user towards the construction of the rom every time the user wants
In added, being the changes in the rom, also allows to enable tweaks as ccache to build the rom

@Pachset2: Fix some stuff that break the tool.
@Pachset3: Add GUI, don't delete Command Line (CL).
@Pachset4: Delete CL, the GUI work; new GUI; Ability to uninstall; Auto Chmod; Switch to xTerm, works better then gnome-terminal.
@Pachset5: Re-Add Build.sh; Add caf devices; Fix th readme.
@Pachset6: Remove OFFICIAL DEV recognition, fix build.sh for ABT.jar, improve ABT(new features, new UI).
Thanks to @akhilnarang now we use time mka instead of time make, now the build is faster

Change-Id: Ia738775b984442630446935822b2948b1a2651c9